TrabalhoSala/requirements.txt
fastapi
uvicorn
plyer
pydantic
heapq
jinja2
sse-starlette
